disabling modification of UIWindow.windowLevel
Modification of UIWindow.windowLevel affects how UIAlerts behave. When UIAlertView is attempted to be shown from left/right viewController and that viewController is "closed", so is the UIAlertView, which is only presented when the left/right viewController is being "opened" again. Also, there's a z-order of rootViewController and status bar is being inverted.